requirements.txt
python-telegram-bot==20.7
youtube-transcript-api
ollama
numpy
langdetect